# api/openapi/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CommentCreateRequest:
    properties:
      content:
        maxLength: 1000
        minLength: 1
        type: string
      parent_id:
        type: integer
    required:
    - content
    type: object
  dto.CommentUpdateRequest:
    properties:
      content:
        maxLength: 1000
        minLength: 1
        type: string
    required:
    - content
    type: object
  dto.LoginRequest:
    properties:
      password:
        maxLength: 255
        minLength: 6
        type: string
      username:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - password
    - username
    type: object
  dto.RegisterRequest:
    properties:
      avatar:
        maxLength: 500
        type: string
      background_image:
        maxLength: 500
        type: string
      password:
        maxLength: 255
        minLength: 6
        type: string
      user_role:
        enum:
        - user
        - admin
        type: string
      username:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - password
    - username
    type: object
...
  dto.TokenData:
    properties:
      expires_in:
        type: integer
      token:
        type: string
      token_type:
        type: string
      user:
        $ref: '#/definitions/dto.UserInfo'
    type: object
  dto.UserInfo:
    properties:
      avatar:
        type: string
      background_image:
        type: string
      follow_count:
        type: integer
      follower_count:
        type: integer
      id:
        type: integer
      user_name:
        type: string
      user_role:
        type: string
    type: object
...
  response.ErrorInfo:
    properties:
      code:
        type: integer
      message:
        type: string
      type:
        type: string
    type: object
  response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/response.ErrorInfo'
    type: object
  response.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: 127.0.0.1:8000
info:
  contact:
    email: support@vida.com
    name: API Support
  description: 视频分享平台 API 服务
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Vida-Go API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: 用户登录获取 JWT Token
      parameters:
      - description: 登录信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.TokenData'
              type: object
        "400":
          description: 请求参数无效
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 用户名或密码错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: 用户登录
      tags:
      - 认证
  /auth/logout:
    post:
      description: 用户登出（当前仅返回成功）
      produces:
      - application/json
      responses:
        "200":
          description: 登出成功
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 用户登出
      tags:
      - 认证
  /auth/me:
    get:
      description: 获取当前登录用户的详细信息
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserInfo'
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取当前用户信息
      tags:
      - 认证
  /auth/register:
    post:
      consumes:
      - application/json
      description: 注册新用户账号
      parameters:
      - description: 注册信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 注册成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserInfo'
              type: object
        "400":
          description: 请求参数无效
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: 用户注册
      tags:
      - 认证
  /comments/{id}:
    delete:
      description: 删除指定评论
      parameters:
      - description: 评论ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: 评论不存在
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 删除评论
      tags:
      - 评论
    put:
      consumes:
      - application/json
      description: 更新指定评论的内容
      parameters:
      - description: 评论ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新内容
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CommentUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: 评论不存在
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 更新评论
      tags:
      - 评论
  /comments/{id}/replies:
    get:
      description: 获取指定评论的回复列表
      parameters:
      - description: 评论ID
        in: path
        name: id
        required: true
        type: integer
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取评论回复列表
      tags:
      - 评论
  /comments/{video_id}:
    post:
      consumes:
      - application/json
      description: 对指定视频发表评论
      parameters:
      - description: 视频ID
        in: path
        name: video_id
        required: true
        type: integer
      - description: 评论内容
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CommentCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 发表成功
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: 视频不存在
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 发表评论
      tags:
      - 评论
  /comments/my/list:
    get:
      description: 获取当前用户发表的评论列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取我的评论列表
      tags:
      - 评论
  /comments/video/{video_id}:
    get:
      description: 获取指定视频的评论列表
      parameters:
      - description: 视频ID
        in: path
        name: video_id
        required: true
        type: integer
      - description: 父评论ID
        in: query
        name: parent_id
        type: integer
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取视频评论列表
      tags:
      - 评论
...
securityDefinitions:
  BearerAuth:
    description: '输入格式: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"